Primary-Key dup-check, improve error messages - Also refactored NewTable functions into separate file
